feat: Add Join2Create system functionality
The code changes include adding the necessary database column `join2CreateChannelId` to the `guilds` model and updating the translation files to support the Join2Create system. This commit adds the functionality for setting up the Join2Create system, including a command to specify the channel for Join2Create messages and corresponding activation and deactivation replies.

Ref: Join2Create #57